@@ -2,15 +2,19 @@
   "name": "spellchecker-cli",
   "version": "4.7.0",
   "description": "A command-line tool for spellchecking files.",
-  "main": "index.js",
+  "main": "build/index.js",
+  "files": [
+    "build"
+  ],
   "bin": {
-    "spellchecker": "./index.js"
+    "spellchecker": "./build/index.js"
   },
   "scripts": {
     "ci": "yarn spellchecker && yarn lint && yarn test",
-    "lint": "eslint \"**/*.js\"",
-    "spellchecker": "node ./index.js",
-    "test": "mocha"
+    "lint": "eslint \"**/*.js\" \"**/*.ts\"",
+    "spellchecker": "ts-node ./index.ts",
+    "test": "tsc && mocha",
+    "prepack": "tsc"
   },
   "author": "Thomas Broadley <buriedunderbooks@hotmail.com>",
   "license": "MIT",
@@ -30,7 +34,7 @@
     "dictionary-en-za": "^2.0.4",
     "dictionary-vi": "^2.1.3",
     "fs-extra": "^6.0.1",
-    "globby": "^8.0.2",
+    "globby": "^11.0.3",
     "js-yaml": "^3.14.1",
     "jsonc": "^2.0.0",
     "junit-report-builder": "^1.3.3",
@@ -48,14 +52,20 @@
     "toml": "^3.0.0",
     "unist-util-visit": "^1.4.1",
     "vfile": "^3.0.1",
-    "vfile-reporter": "^5.1.2"
+    "vfile-reporter": "^6.0.0"
   },
   "devDependencies": {
+    "@types/fs-extra": "^9.0.11",
+    "@types/node": "^15.12.2",
+    "@typescript-eslint/eslint-plugin": "^4.26.1",
+    "@typescript-eslint/parser": "^4.26.1",
     "chai": "^4.3.4",
-    "eslint": "^5.16.0",
+    "eslint": "^7.28.0",
     "eslint-config-airbnb-base": "^13.2.0",
     "eslint-plugin-import": "^2.23.3",
     "mocha": "^8.4.0",
-    "mocha.parallel": "0.15.5"
+    "mocha.parallel": "0.15.5",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
   }
 }
